Fixed build error and moved test files to correct module
